Added new button for data generating and comments to view-functions (homework_06).
