+++ homework_06/Dockerfile
@@ -1,9 +1,3 @@
-# -------------------------------------------- #
-# Docker-файл для сборки контейнера приложения;
-# 0) Взяли из заготовки;
-# -------------------------------------------- #
-
-# ---------------------- Step 0 ---------------------- #
 FROM tiangolo/uwsgi-nginx-flask:python3.10
 
 COPY requirements.txt /app/requirements.txt
