--- homework_06/app/requirements.txt
@@ -1,89 +1,0 @@
-aiohttp==3.8.3
-aiosignal==1.3.1
-alembic==1.9.1
-anyio==3.6.2
-asn1crypto==1.5.1
-async-timeout==4.0.2
-asyncio==3.4.3
-asyncpg==0.27.0
-attrs==22.1.0
-CacheControl==0.12.11
-cachy==0.3.0
-certifi==2022.9.14
-cffi==1.15.1
-charset-normalizer==2.1.1
-cleo==1.0.0a5
-click==8.1.3
-crashtest==0.3.1
-cryptography==38.0.1
-distlib==0.3.6
-docker==5.0.3
-dulwich==0.20.50
-Faker==13.15.0
-fastapi==0.85.1
-filelock==3.8.0
-Flask==2.2.2
-Flask-Migrate==4.0.1
-Flask-SQLAlchemy==3.0.2
-Flask-WTF==1.0.1
-frozenlist==1.3.3
-greenlet==2.0.1
-h11==0.14.0
-html5lib==1.1
-idna==3.4
-iniconfig==1.1.1
-itsdangerous==2.1.2
-jaraco.classes==3.2.3
-jeepney==0.8.0
-Jinja2==3.1.2
-jsonschema==4.16.0
-keyring==23.9.3
-lockfile==0.12.2
-Mako==1.2.4
-MarkupSafe==2.1.1
-more-itertools==9.0.0
-msgpack==1.0.4
-multidict==6.0.3
-mypy==0.991
-mypy-extensions==0.4.3
-packaging==21.3
-pexpect==4.8.0
-pg8000==1.29.4
-pkginfo==1.8.3
-platformdirs==2.5.2
-pluggy==1.0.0
-poetry==1.2.2
-poetry-core==1.3.2
-poetry-plugin-export==1.1.2
-psycopg2-binary==2.9.5
-ptyprocess==0.7.0
-py==1.11.0
-pycparser==2.21
-pydantic==1.10.2
-pylev==1.4.0
-pyparsing==3.0.9
-pyrsistent==0.18.1
-pytest==7.1.2
-pytest-asyncio==0.18.3
-python-dateutil==2.8.2
-requests==2.28.1
-requests-toolbelt==0.9.1
-scramp==1.4.4
-SecretStorage==3.3.3
-shellingham==1.5.0
-six==1.16.0
-sniffio==1.3.0
-SQLAlchemy==1.4.45
-sqlalchemy-stubs==0.4
-starlette==0.20.4
-tomli==2.0.1
-tomlkit==0.11.6
-typing_extensions==4.4.0
-urllib3==1.26.12
-uvicorn==0.19.0
-virtualenv==20.16.6
-webencodings==0.5.1
-websocket-client==1.4.1
-Werkzeug==2.2.2
-WTForms==3.0.1
-yarl==1.8.2
