+++ homework_06/app/templates/index.html
@@ -4,7 +4,15 @@
 {% endblock %}
 {% block body %}
     <div class="h2">Main info</div>
-    <a class="text-decoration-none my-3"
-        href="{{ url_for('users_app.user_list') }}"
-    >Go to user list</a>
+    <button type="button"
+            class="btn btn-outline-warning my-3"
+            onclick="GenerateNewBill()"
+            id="Generate_new_bills">
+        Generate bills
+    </button>
+    <script>
+        function GenerateNewBill() {
+            window.location.href = "{{ url_for('users_app.reload_users') }}"
+        }
+    </script>
 {% endblock %}
